+++ src/meltano/api/controllers/repos.py
@@ -1,19 +1,23 @@
 import base64
 import json
 import os
-import subprocess
-import sys
 from pathlib import Path
 from os.path import join
 
 import markdown
-import pkg_resources
-from flask import Blueprint, jsonify
+from flask import Blueprint, jsonify, request
 
+from meltano.api.security import api_auth_required
 from .m5o_file_parser import MeltanoAnalysisFileParser, MeltanoAnalysisFileParserError
 
 reposBP = Blueprint("repos", __name__, url_prefix="/repos")
 meltano_model_path = join(os.getcwd(), "model")
+
+
+@reposBP.before_request
+@api_auth_required
+def before_request():
+    pass
 
 
 @reposBP.route("/", methods=["GET"])
@@ -25,7 +29,13 @@ def index():
         for f in os.listdir(meltano_model_path)
         if os.path.isfile(os.path.join(meltano_model_path, f))
     ]
-    sortedM5oFiles = {"documents": [], "tables": [], "models": [], "dashboards": []}
+    sortedM5oFiles = {
+        "documents": [],
+        "tables": [],
+        "models": [],
+        "dashboards": [],
+        "reports": [],
+    }
     onlydocs = Path(meltano_model_path).parent.glob("*.md")
     for d in onlydocs:
         file_dict = {"path": str(d), "abs": str(d), "visual": str(d.name)}
@@ -52,6 +62,8 @@ def index():
             sortedM5oFiles["models"].append(file_dict)
         if ext == ".dashboard":
             sortedM5oFiles["dashboards"].append(file_dict)
+        if ext == ".report":
+            sortedM5oFiles["reports"].append(file_dict)
 
     return jsonify(sortedM5oFiles)